refactor: reorganize SwarmSwIM package into behaviors/ and analysis/
Move navigation scripts into SwarmSwIM/behaviors/ and plotting scripts
into SwarmSwIM/analysis/ to separate concerns within the package.
Move example2Danimation.py to top-level examples/ directory.

Structural changes:
- SwarmSwIM/behaviors/  flocking_nav, flocking_ang_slow, aggregation_nav,
                        agg_nav_v2, regatta_nav
- SwarmSwIM/analysis/   plot10, plotsingle, plotaggregate
- examples/             example2Danimation.py

Import fixes:
- Nav scripts: relative imports .sensors/.sail_extension -> ..sensors/..sail_extension
- Nav scripts: base = Path(__file__).parent -> .parent.parent (one level deeper)
- plot10.py: BASE_DIR parent depth +1 (analysis/ -> SwarmSwIM/ -> repo root)
- example2Danimation.py: relative imports -> absolute SwarmSwIM.* imports

Shell scripts updated to python3 -m SwarmSwIM.behaviors.*
README and results/ READMEs updated with new paths.
Public API (from SwarmSwIM import Simulator, Agent, Plotter) unchanged.
