+++ SwarmSwIM/analysis/__init__.py
@@ -1,0 +1,5 @@
+# Post-simulation analysis and plotting scripts for SwarmSwIM.
+# Run any script directly from the repo root:
+#   python -m SwarmSwIM.analysis.plot10
+#   python -m SwarmSwIM.analysis.plotsingle
+#   python -m SwarmSwIM.analysis.plotaggregate
