+++ SwarmSwIM/analysis/plotaggregate.py
