+++ SwarmSwIM/analysis/plotsingle.py
