+++ SwarmSwIM/behaviors/__init__.py
@@ -1,0 +1,7 @@
+# Navigation behavior scripts for SwarmSwIM.
+# Run any behavior directly from the repo root:
+#   python -m SwarmSwIM.behaviors.flocking_nav
+#   python -m SwarmSwIM.behaviors.aggregation_nav
+#   python -m SwarmSwIM.behaviors.agg_nav_v2
+#   python -m SwarmSwIM.behaviors.flocking_ang_slow
+#   python -m SwarmSwIM.behaviors.regatta_nav
